requirements.txt
sanic==0.5.4
sanic-jinja2==0.5.3
motor==1.1
Sanic-WTF==0.5.0
python-slugify==1.2.4
